+++ src/Lifecycle/Deactivator.php
@@ -4,15 +4,15 @@ namespace WPPluginBoilerplate\Lifecycle;
 
 use WPPluginBoilerplate\Plugin;
 
-class Deactivator
-{
-	public static function deactivate(): void
-	{
-		wp_clear_scheduled_hook(
-			Plugin::cron_hook('cron')
-		);
+class Deactivator {
 
-		flush_rewrite_rules();
+	public static function deactivate(): void {
+
+		/**
+		 * Intentionally do NOT delete options here.
+		 * Only runtime cleanup should live here.
+		 */
+
+		do_action(  Plugin::prefix() . 'deactivated' );
 	}
 }
-